+++ redgifs/ui/src/main/java/com/example/ui/screens/explorer/tab/saved/tab/collection/SavedCollectionName.kt
@@ -103,8 +103,8 @@ class ScreenCollectionName(val collectionName: String) : Screen {
             savedRed.collections.selectedCollection.value = null
         }
 
-        LaunchedEffect(SavedCollectionTab.columnSelect.column.intValue) {
-            vm.likedHost.columns = SavedCollectionTab.columnSelect.column.intValue
+        LaunchedEffect(SavedCollectionTab.columnSelect.column) {
+            vm.likedHost.columns = SavedCollectionTab.columnSelect.column
         }
 
         /* ---------- /Диалог ---------- */
